First stable release.

+++ src/InputData.java
@@ -1,10 +1,10 @@
 public class InputData {
-    private int a; // lower limit
-    private int b; // upper limit
+    private double a; // lower limit
+    private double b; // upper limit
     private double precision;
     private Functions function;
 
-    InputData(Functions function, int a, int b, double precision) {
+    InputData(Functions function, double a, double b, double precision) {
         this.a = a;
         this.b = b;
         this.precision = precision;
@@ -14,12 +14,12 @@ public class InputData {
     public Functions getFunction() {
         return function;
     }
-    
-    public int getUpperLimit() {
+
+    public double getUpperLimit() {
         return b;
     }
 
-    public int getLowerLimit() {
+    public double getLowerLimit() {
         return a;
     }
